Extract admin-only ProductVariantsGenerateType
