+++ src/Sylius/Bundle/AdminBundle/Resources/config/services/twig/component.xml
@@ -267,7 +267,7 @@
 
         <service id="Sylius\Bundle\AdminBundle\Twig\Component\Product\GenerateProductVariantsFormComponent">
             <argument type="service" id="form.factory" />
-            <argument>Sylius\Bundle\ProductBundle\Form\Type\ProductGenerateVariantsType</argument>
+            <argument>Sylius\Bundle\AdminBundle\Form\Type\ProductGenerateVariantsType</argument>
         </service>
 
         <service
